Broke up physics file, QOL physics interface updates
