+++ include/interface/simulation.h
@@ -27,8 +27,8 @@ size_t sim_get_nearby(SIMULATION *sim, COLLISION **dest, vec3 pos,
 size_t save_sim_state(SIMULATION *sim, SIM_STATE **state);
 void restore_sim_state(SIMULATION *sim, SIM_STATE *state, size_t state_size);
 void free_sim_state(SIM_STATE *state, size_t num_ents);
-void impulse_resolution(SIMULATION *sim, COLLISION col, vec3 a_dest_vel,
-                        vec3 b_dest_vel);
+void impulse_collision(COL_ARGS *a_args, COL_ARGS *b_args, vec3 p_dir,
+                       vec3 p_loc, vec3 gravity);
 
 void draw_oct_tree(MODEL *cube, OCT_TREE *tree, vec3 pos, float scale,
                    unsigned int shader, size_t offset, int depth);
