+++ include/interface/structs/simulation_str.h
@@ -90,6 +90,7 @@ typedef struct collision_list {
 
 typedef struct collision_args {
   float *vel_dest;
+  float *ang_vel_dest;
   float *velocity;
   float *ang_velocity;
   size_t collider;
