+++ src/physics/simulation.c
@@ -381,157 +381,6 @@ void integrate_sim(SIMULATION *sim, vec3 origin, float range) {
   }
 }
 
-// Default integrate_ent callback
-void integrate_ent(ENTITY *ent, vec3 forces) {
-  if (ent->num_cons) {
-    apply_constraints(ent, ent->p_cons, ent->num_cons, forces);
-  }
-  featherstone_abm(ent, forces);
-
-  float delta = 0.0;
-  for (int cur_bone = 0; cur_bone < ent->model->num_bones; cur_bone++) {
-    int cur_col = ent->model->bone_collider_links[cur_bone];
-    if (cur_col == -1) {
-      continue;
-    }
-
-    int collider_root_bone = ent->model->collider_bone_links[cur_col];
-    if (collider_root_bone == cur_bone) {
-      // Integrate acceleration of zero joints
-      for (size_t i = 0; i < ent->np_data[cur_col].num_z_joints; i++) {
-        size_t cur_zj = ent->np_data[cur_col].zero_joint_offset + i;
-        delta = ent->zj_data[cur_zj].accel_angle * DELTA_TIME;
-        ent->zj_data[cur_zj].vel_angle *= 0.999;
-        ent->zj_data[cur_zj].vel_angle += delta;
-        remove_noise(ent->zj_data[cur_zj].vel_angle, ZERO_THRESHOLD);
-      }
-      // Integrate acceleration of main joint
-      delta = ent->np_data[cur_col].accel_angle * DELTA_TIME;
-      ent->np_data[cur_col].vel_angle *= 0.999;
-      ent->np_data[cur_col].vel_angle += delta;
-      remove_noise(ent->np_data[cur_col].vel_angle, ZERO_THRESHOLD);
-
-      ///*
-      //fprintf(stderr, "a: %f, %f, %f\n", ent->np_data[cur_col].a[0],
-      //        ent->np_data[cur_col].a[1], ent->np_data[cur_col].a[2]);
-      vec3 delta_vec3 = GLM_VEC3_ZERO_INIT;
-      glm_vec3_scale(ent->np_data[cur_col].a, DELTA_TIME, delta_vec3);
-      //fprintf(stderr, "d_v: %f, %f, %f\n", delta_vec3[0], delta_vec3[1],
-      //        delta_vec3[2]);
-      //fprintf(stderr, "v before: %f, %f, %f\n", ent->np_data[cur_col].v[0],
-      //        ent->np_data[cur_col].v[1], ent->np_data[cur_col].v[2]);
-      glm_vec3_scale(ent->np_data[cur_col].v, 0.999,
-                     ent->np_data[cur_col].v);
-      glm_vec3_add(ent->np_data[cur_col].v, delta_vec3,
-                   ent->np_data[cur_col].v);
-      vec3_remove_noise(ent->np_data[cur_col].v, ZERO_THRESHOLD);
-      //fprintf(stderr, "v after: %f, %f, %f\n", ent->np_data[cur_col].v[0],
-      //        ent->np_data[cur_col].v[1], ent->np_data[cur_col].v[2]);
-
-      //fprintf(stderr, "ang_a: %f, %f, %f\n", ent->np_data[cur_col].ang_a[0],
-      //        ent->np_data[cur_col].ang_a[1], ent->np_data[cur_col].ang_a[2]);
-      glm_vec3_scale(ent->np_data[cur_col].ang_a, DELTA_TIME, delta_vec3);
-      //fprintf(stderr, "d_v: %f, %f, %f\n", delta_vec3[0], delta_vec3[1],
-      //        delta_vec3[2]);
-      //fprintf(stderr, "ang_v before: %f, %f, %f\n",
-      //        ent->np_data[cur_col].ang_v[0], ent->np_data[cur_col].ang_v[1],
-      //        ent->np_data[cur_col].ang_v[2]);
-      glm_vec3_scale(ent->np_data[cur_col].ang_v, 0.999,
-                     ent->np_data[cur_col].ang_v);
-      glm_vec3_add(ent->np_data[cur_col].ang_v, delta_vec3,
-                   ent->np_data[cur_col].ang_v);
-      vec3_remove_noise(ent->np_data[cur_col].ang_v, ZERO_THRESHOLD);
-      //fprintf(stderr, "ang_v after: %f, %f, %f\n\n",
-      //        ent->np_data[cur_col].ang_v[0], ent->np_data[cur_col].ang_v[1],
-      //        ent->np_data[cur_col].ang_v[2]);
-      //*/
-
-      // Integrate velocity of zero joints
-      for (size_t i = 0; i < ent->np_data[cur_col].num_z_joints; i++) {
-        size_t cur_zj = ent->np_data[cur_col].zero_joint_offset + i;
-        delta = ent->zj_data[cur_zj].vel_angle * DELTA_TIME;
-        ent->zj_data[cur_zj].joint_angle += delta;
-      }
-
-      // Integrate velocity of main joint
-      delta = ent->np_data[cur_col].vel_angle * DELTA_TIME;
-      ent->np_data[cur_col].joint_angle += delta;
-      glm_vec3_scale(ent->np_data[cur_col].v, DELTA_TIME, delta_vec3);
-      glm_translate(ent->bone_mats[cur_bone][LOCATION], delta_vec3);
-
-      vec3 delta_rot = GLM_VEC3_ZERO_INIT;
-      glm_vec3_scale(ent->np_data[cur_col].ang_v, DELTA_TIME, delta_rot);
-      versor rot_quat = GLM_QUAT_IDENTITY_INIT;
-      glm_quatv(rot_quat, glm_vec3_norm(delta_rot), delta_rot);
-      glm_quat_normalize(rot_quat);
-      versor temp_quat = GLM_QUAT_IDENTITY_INIT;
-      glm_mat4_quat(ent->bone_mats[cur_bone][ROTATION], temp_quat);
-      glm_quat_normalize(temp_quat);
-      glm_quat_mul(rot_quat, temp_quat, temp_quat);
-      glm_quat_normalize(temp_quat);
-      glm_quat_mat4(temp_quat, ent->bone_mats[cur_bone][ROTATION]);
-
-      // Combine rotation, location and scale into final bone matrix
-      vec3 temp = GLM_VEC3_ZERO_INIT;
-      mat4 from_center = GLM_MAT4_IDENTITY_INIT;
-      mat4 to_center = GLM_MAT4_IDENTITY_INIT;
-      if (ent->model->colliders[cur_col].type == SPHERE) {
-        glm_vec3_copy(ent->model->colliders[cur_col].data.center, temp);
-      } else {
-        glm_vec3_copy(ent->model->colliders[cur_col].data.center_of_mass,
-                      temp);
-      }
-      glm_translate(to_center, temp);
-      glm_vec3_negate(temp);
-      glm_translate(from_center, temp);
-
-      glm_mat4_identity(ent->final_b_mats[cur_bone]);
-      glm_mat4_mul(from_center, ent->final_b_mats[cur_bone],
-                   ent->final_b_mats[cur_bone]);
-      glm_mat4_mul(ent->bone_mats[cur_bone][SCALE],
-                   ent->final_b_mats[cur_bone],
-                   ent->final_b_mats[cur_bone]);
-      glm_mat4_mul(ent->bone_mats[cur_bone][ROTATION],
-                   ent->final_b_mats[cur_bone],
-                   ent->final_b_mats[cur_bone]);
-      glm_mat4_mul(to_center, ent->final_b_mats[cur_bone],
-                   ent->final_b_mats[cur_bone]);
-      glm_mat4_mul(ent->bone_mats[cur_bone][LOCATION],
-                   ent->final_b_mats[cur_bone],
-                   ent->final_b_mats[cur_bone]);
-
-      // "Anchor" revolute joints so they are not affected by translational
-      // drift
-      int parent_bone = ent->model->bones[cur_bone].parent;
-      vec3 anchor = GLM_VEC3_ZERO_INIT;
-      if (parent_bone != -1 &&
-          ent->np_data[cur_col].joint_type != JOINT_PRISMATIC) {
-        vec3 base_loc = GLM_VEC3_ZERO_INIT;
-        glm_mat4_mulv3(ent->final_b_mats[cur_bone],
-                       ent->model->bones[cur_bone].base, 1.0, base_loc);
-        vec3 p_head_loc = GLM_VEC3_ZERO_INIT;
-        glm_mat4_mulv3(ent->final_b_mats[parent_bone],
-                       ent->model->bones[parent_bone].head, 1.0,
-                       p_head_loc);
-        glm_vec3_sub(p_head_loc, base_loc, anchor);
-      }
-      mat4 anchor_mat = GLM_MAT4_IDENTITY_INIT;
-      glm_translate(anchor_mat, anchor);
-      glm_mat4_mul(anchor_mat, ent->final_b_mats[cur_bone],
-                   ent->final_b_mats[cur_bone]);
-      glm_mat4_mul(anchor_mat, ent->bone_mats[cur_bone][LOCATION],
-                   ent->bone_mats[cur_bone][LOCATION]);
-    } else {
-      glm_mat4_copy(ent->final_b_mats[collider_root_bone],
-                    ent->final_b_mats[cur_bone]);
-    }
-  }
-
-  for (size_t j = 0; j < ent->model->num_colliders; j++) {
-    glm_vec3_zero(ent->np_data[j].e_force);
-  }
-}
-
 size_t get_sim_collisions(SIMULATION *sim, COLLISION **dest, vec3 origin,
                           float range, int get_col_info) {
   pthread_mutex_t col_lock;
@@ -720,51 +569,6 @@ size_t sim_get_nearby(SIMULATION *sim, COLLISION **dest, vec3 pos,
   free(collisions);
 
   return buf_len;
-}
-
-void impulse_resolution(SIMULATION *sim, COLLISION col, vec3 a_dest_vel,
-                        vec3 b_dest_vel) {
-  COL_ARGS a_args;
-  a_args.vel_dest = a_dest_vel;
-  a_args.velocity = col.a_ent->np_data[col.a_offset].v;
-  a_args.ang_velocity = col.a_ent->np_data[col.a_offset].ang_v;
-  a_args.collider = col.a_offset;
-  a_args.type = col.a_ent->type;
-  int bone = col.a_ent->model->collider_bone_links[col.a_offset];
-  glm_mat4_quat(col.a_ent->bone_mats[bone][ROTATION], a_args.rotation);
-  a_args.inv_mass = col.a_ent->np_data[col.a_offset].inv_mass;
-  if (a_args.inv_mass) {
-    calc_inertia_tensor(col.a_ent, col.a_offset, a_args.inv_mass,
-                        a_args.inv_inertia);
-    glm_mat4_inv(a_args.inv_inertia, a_args.inv_inertia);
-  } else {
-    glm_mat4_zero(a_args.inv_inertia);
-  }
-  glm_mat4_mulv3(col.a_ent->final_b_mats[bone],
-                 col.a_ent->model->bones[bone].base, 1.0,
-                 a_args.center_of_rotation);
-
-  COL_ARGS b_args;
-  b_args.vel_dest = b_dest_vel;
-  b_args.velocity = col.b_ent->np_data[col.b_offset].v;
-  b_args.ang_velocity = col.b_ent->np_data[col.b_offset].ang_v;
-  b_args.collider = col.b_offset;
-  b_args.type = col.b_ent->type;
-  bone = col.b_ent->model->collider_bone_links[col.b_offset];
-  glm_mat4_quat(col.b_ent->bone_mats[bone][ROTATION], b_args.rotation);
-  b_args.inv_mass = col.b_ent->np_data[col.b_offset].inv_mass;
-  if (b_args.inv_mass) {
-    calc_inertia_tensor(col.b_ent, col.b_offset, b_args.inv_mass,
-                        b_args.inv_inertia);
-    glm_mat4_inv(b_args.inv_inertia, b_args.inv_inertia);
-  } else {
-    glm_mat4_zero(b_args.inv_inertia);
-  }
-  glm_mat4_mulv3(col.b_ent->final_b_mats[bone],
-                 col.b_ent->model->bones[bone].base, 1.0,
-                 b_args.center_of_rotation);
-
-  solve_collision(&a_args, &b_args, col.col_dir, col.col_point, sim->forces);
 }
 
 // ============================== MISC HELPERS ===============================
